Update to handle One Platform discovery
In addition to some invisible stuff with the codegen, this commit
also lets users specify their own separators with resource names.

These custom separators are necessary because some resource names
fundamentally contain slashes that are not indicated by discovery
docs.

To use a custom separator, prefix the resource name with something
that is not a letter or a digit, and that prefix will be used to
split the rest of the argument. If the first character is a letter
or digit, the previous '/' will be used as a default.